+++ Myscelium/rust/src/common/functions/python_functions.rs
@@ -21,6 +21,48 @@ use crate::common::enhanced_buffer::utilities::Command;
 
 use serde_json::{json, Value};
 
+use serde_json::Value as JsonValue;
+
+/// Converts a JSON value to its corresponding Python object.
+///
+/// This helper function takes in a JSON value and recursively converts it to the corresponding Python object.
+/// This can be useful for translating between Rust and Python data structures.
+///
+/// # Parameters
+///
+/// - `py`: The Python interpreter.
+/// - `value`: The JSON value to convert.
+///
+/// # Returns
+///
+/// Returns the converted Python object.
+pub fn translate_value_to_py(py: Python<'_>, value: JsonValue) -> PyResult<PyObject> {
+    // Convert the JSON value to the appropriate Python object
+    match value {
+        JsonValue::Null => Ok(py.None()),
+        JsonValue::Bool(b) => Ok(b.into_py(py)),
+        JsonValue::Number(num) => Ok(num.as_f64().unwrap().into_py(py)),
+        JsonValue::String(s) => Ok(s.into_py(py)),
+        JsonValue::Array(arr) => {
+            let py_list = PyList::empty(py);
+            for item in arr {
+                let py_item = translate_value_to_py(py, item)?;
+                py_list.append(py_item)?;
+            }
+            Ok(py_list.into())
+        },
+        JsonValue::Object(obj) => {
+            let py_dict: &PyDict = PyDict::new(py);
+            for (k, v) in obj {
+                let py_key = k.into_py(py);
+                let py_value = translate_value_to_py(py, v)?;
+                py_dict.set_item(py_key, py_value)?;
+            }
+            Ok(py_dict.into())
+        },
+    }
+}
+
 pub fn extract_arg_types(arg: &PyAny) -> PyResult<Value> {
     if let Ok(arg_dict) = arg.downcast::<PyDict>() {
         // If the argument is a dictionary, recursively extract the argument types