+++ Myscelium/rust/src/lib.rs
@@ -120,6 +120,8 @@ fn myscelium_engine(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(client_send, m)?)?;
     m.add_function(wrap_pyfunction!(set_client_uid, m)?)?;
     m.add_function(wrap_pyfunction!(set_socket_client_log_level, m)?)?;
+    m.add_function(wrap_pyfunction!(get_socket_client_available_handlers, m)?)?;
+
     // m.add_function(wrap_pyfunction!(registry_client_logs_handler, m)?)?;
 
     Ok(())