+++ Myscelium/rust/src/socket_client/mod.rs
@@ -1,4 +1,5 @@
 pub mod client_logger;
+mod functions;
 pub mod scheduler;
 pub mod socket_client;
 pub mod transposer;